requirements.txt
opencv-python-headless
pytesseract
Pillow
selenium
pytest
webdriver_manager